Fixes and additions to list.
Fixed a bug in priority lists, where adding an element to the head or tail of the list wouldn't set the priority.

Added list_free, which frees the memory associated with each element of a list.
Added list_split, which splits a list into two at the given element.
Added list_remove, which searches a list for an element and removes it.




git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@412 a4d8336d-3463-0410-8bba-c098c45d37a8