For work_queue:
When the master wants to send an input file to a worker, it sends the following command first:

< "put %s %d %o\n",filename,&length,&mode >

In this revision, I changed the command to:

< "put %s %d %o %d\n",filename,&length,&mode, &cacheable > (added argument 'cacheable')

So the worker can delete uncacheable files upon completion of a task.

(This solves the problem that a SAND user has pointed out, which is the worker accumulating input files from previous tasks.)

For list.c
the list_pop_head and list_pop_tail function now would check if the function parameter (struct list *) itself is NULL at the beginning.



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@622 a4d8336d-3463-0410-8bba-c098c45d37a8