1. added -a option to allow worker to work under auto mode which asks for work from our default catalog server(chirp.cse.nd.edu);
what's going on with -a option: 

the worker would download the masters' information (a list of masters) from the catalog server and sort them by priority in descendent order. If a master matches the worker's preference (specified by the -p option, see below), its priority gets promoted. Then the worker would traverse the master list and connect to the first one that it is able to make a successful connection.

2. when -a option is specified, use -p <project> to specify which project this work would prefer to work on. Right now, it only supports one preference. In the future, we might want a list of preferred projects;

3. changed Kevin's LINE_MAX #define in get_line.h to LARGE_LINE_MAX because it conflicts with some include file in link.h 



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@634 a4d8336d-3463-0410-8bba-c098c45d37a8