batch_job: add '=' filenaming syntax to WQ backend
Users may now take advantage of WQ's renaming facilities for input and output
files by specifying the mapping using `=`.  For instance to specify a local
input file `/a/b/c` is to be renamed `foo` on the remote node, specify the
input as `/a/b/c=foo`.  Likewise, to name a remote output file `bar` to
`/d/e/f` on the local system, specify `bar=/d/e/f`.  This means that filenames
cannot contain a `=` otherwise, this new filename parsing will take place.  In
the absence of `=`, the old standard file naming specification will apply.

git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@639 a4d8336d-3463-0410-8bba-c098c45d37a8