work_queue: distinguish between OUTPUT_FAIL and LINK_FAIL
When a task completes, attempt to retrieve the output files.  If an output file
does not exist, then mark the task result as OUTPUT_FAIL.  If the connection
drops in the middle of the retrieval, then mark the task result as LINK_FAIL.
Now, when we go to remove the worker, we check the task result.  If we have
reached OUTPUT_FAIL, then we move the task to the complete list.  Otherwise, we
move it to ready list, to be retried internally by the WQ.

This change now allows Makeflow to continue processing the DAG whenever a task
fails to return its output.  Before this, the WQ would continuously retry the
task internally, and thus Makeflow would not make any progress.

The small change to batch_job allows Makeflow to check the results of a failed
WQ task and reschedule it if the retry flag is specified.

As a side note, we may wish to consider the case where the output failed, but
the worker is still OK to run more tasks.  In that case, removing the worker is
unnecessary.  Right now, if any part of the task execution fails, we always
remove the responsible worker.



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@812 a4d8336d-3463-0410-8bba-c098c45d37a8