pfs_service_bxgrid: add access syscall
This fixes bug where Makeflow performs an access and it succeeds even though
the file does not exist.  This leads to problems when Makeflow dispatches the
job to WorkQueue, even though the file is not really there, and thus WorkQueue
gets stuck, constantly retrying a job while Makeflow spins its wheels waiting
for a task that will never be returned.


git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@966 a4d8336d-3463-0410-8bba-c098c45d37a8