In work queue, take upper-level application time into account when calculating the master capacity.
Instead of keeping track of avg task transfer time and avg task execution time for all the tasks, only do the avg on the number of tasks that equal to the number of workers, which is similar to how we calculate idle time and avg upper-level application time/task.

Restructured some functions and removed debug code (time measuring code, so these code can be found in the previous revision) and useless comments.



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/branches/lyu2/wq-accounting@1271 a4d8336d-3463-0410-8bba-c098c45d37a8