Make swig path configurable
It appears that swig 1.3.29 will cause segfaults with the Python WorkQueue
bindings, but swig 1.3.40 does not.  RHEL5, which is still our default target,
has the older version, while RHEL6, where we do most of our testing has the
later.  Since swig 1.3.40 is in external, always use that when we build to
ensure consistency.



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@1719 a4d8336d-3463-0410-8bba-c098c45d37a8