makeflow: set per-node BATCH_OPTIONS before submissions
The semantics for BATCH_OPTIONS are as follows:

If @BATCH_OPTIONS is set locally to the node, then that is used when submitting
the task.  Otherwise, if BATCH_OPTIONS is set in the DAG or in the environment,
then that is used.

Since batch_job currently only supports global batch_queue options, we must
store the previous BATCH_OPTIONS value before setting, and restore the old
value after we submit.

Next commmit will test that it functions properly.



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/branches/pbui/makeflow-dissertation@1791 a4d8336d-3463-0410-8bba-c098c45d37a8