makeflow: introduce MAKEFLOW node command to denote recursive Makeflow
The MAKEFLOW node command takes two arguments:

    1. DAG: the Makeflow DAG to execute
    2. CWD: the working directory to use when executing the DAG; this is
    optional and defaults to "." if none is given.

    These arguments are used to generate the following node command:

	cd "CWD"; makeflow "DAG"

Example:

    output: sub.makeflow
	    MAKEFLOW sub.makeflow

Note, this is simply a start at getting the mechanics of recursive Makeflow
started.  The issues of propagating configuration and things like that are not
addressed (nor is it my goal).  My main intention is to simply allow Makeflow
to *know* that a recursive call is occurring.  What we do with this information
is up to someone else (GLHF).

With this new information, makeflow -c now does the "right" thing (i.e. it
recursively invokes the clean operation on recursive Makeflows).



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/branches/pbui/makeflow-dissertation@1888 a4d8336d-3463-0410-8bba-c098c45d37a8