added '-P' option to work_queue_pool.
This option is the same as the '-A' option (auto worker pool) except that it does not start workers automatically by itself.
It only updates the distribution decision to the catalog server so that related workers can see.
This is a way to manage Amazon EC2 and Windows Azure workers with worker pool policies for now.



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@1905 a4d8336d-3463-0410-8bba-c098c45d37a8