Merging r1809 from /trunk: ------------------------------------------------------------------------ r1809 | lyu2 | 2012-02-28 18:10:34 -0500 (Tue, 28 Feb 2012) | 9 lines
changed the D_WQ debug messages before the remove_worker() function calls to
D_NOTICE so that the reason why a worker is rejected or removed can be seen
even when the application is not running in the debug mode (i.e. with the -d
option).

made sure that every remove_worker() call is coupled with an explanation
message if appropriate.


------------------------------------------------------------------------

git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/release/3.4.3@1929 a4d8336d-3463-0410-8bba-c098c45d37a8