Merging r1846 from /trunk: ------------------------------------------------------------------------ r1846 | dthain | 2012-03-06 13:05:37 -0500 (Tue, 06 Mar 2012) | 5 lines
Added support for writable memory mapped files, which is required for the CMS software stack.
The I/O channel now keeps a reference-counted list of the memory objects loaded.
These are referred to by pfs_mmap structures in the pfs_mmap table.
On an munmap, the correct map is looked up and flushed through to disk.

------------------------------------------------------------------------

git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/release/3.4.3@1935 a4d8336d-3463-0410-8bba-c098c45d37a8