makeflow: avoid rerunning nodes if missing files were garbage collected
Now when we schedule a node to be re-run, we have check to see if any of our
parents generated inputs that were garbage collected and rerun them to ensure
our task has the necessary inputs.

Also add a minor check to see if a file exists before unlinking it.  This
suppresses an unnecessary warning.



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/branches/pbui/makeflow-dissertation@1986 a4d8336d-3463-0410-8bba-c098c45d37a8