So after about 6 hours of bug hunting...
pwrite can return 0 on debian 7.0 (linux 3.2.0-2-amd64).

This is IMO a stupid thing to do (either write something or return a legit
error). It caused tracer.c to think it actually wrote the data when it did not.
So, I fixed the result checking.

[For those curious, it's also not unheardof behavior. On FreeBSD 8.X, I have been
able to get a similar 0 return when writing directly to a disk, e.g.:

open("/dev/ada0", O_WRONLY);

and then when you write past the end of the drive, FreeBSD for some reason returns
0 from write rather than telling you that you've exceeded the size of the medium.
Stupidity all around!]


git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@2169 a4d8336d-3463-0410-8bba-c098c45d37a8