Improvements to version printing.
(a) Version printing is unified into dttools/src/cctools.c so
    all tools are consistent in how they print the version.

(b) The release date is automatically generated if not
    hardcoded so autobuild versions make sense.

(c) Configure arguments are added to the version information
    to assist in debugging (for us and users).

(d) There is now a "release type" (for lack of a better word)
    which tags the built version. The default /trunk type
    is "TRUNK". A release like 3.5.2 would probably use
    "RELEASE". Maybe autobuild will use "AUTOBUILD".

All CCTools executables now print for D_DEBUG the version (and configure
flags). parrot_run has always done this.


git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@2196 a4d8336d-3463-0410-8bba-c098c45d37a8