Segfault fix:
This commit includes the main fix for the segfault reported by Lee-Ping. This should also fix the segfault reported by Badi's group recently.

The cause for the segfaut is: if the master terminates while a worker is running a task for it, the worker would disconnect the master, kill the running task and mark its task_status as cancelled. When a new master comes in and uses the same worker, the worker would report the killed task's result (which is result with exit code 15) to the master. However, the master might have just connected the worker and haven't sent any task to it. This corrupts the state of the master and caused the segfault.

This fix entirely removes the TASK_CANCELLED state on the worker and correctified/simplified the task state transitions in work_queue_worker. Now the worker only has two possible task status: TASK_NONE, TASK_RUNNING. If the task status is TASK_NONE, there should be no task processes waiting to be reaped and the stdout buffer and other task related resources should be cleaned up. 

So, now in kill_task() function, we call wait_task_process(), which renamed and modified from collect_task_info(), to wait on the task info and clean up task related resources (currently stdout buffer/file).

Other fixes:

1. Removed the task_status != TASK_CANCELLED check in do_unlink and changed it to: if the path already does not exist, then return success.
2. set stdout_buffer to NULL after freeing it (in abort_worker()). This prevents segfault from stdout_buffer being used after it's free'ed. e.g. in cleaning up task resources in abort_worker().
3. updated debug messages to generate more useful messages in work_queue.c and work_queue_worker.c.




git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@2435 a4d8336d-3463-0410-8bba-c098c45d37a8