Modified makeflow to separate DAG parsing from the preparatory steps for running on a given batch system.
Also enabled remote file renaming on some systems that support it. For example:

c.out=out: a.in=in1 b.in=in2 ./myprog
	./myprog in1 in2 > out

will create a file named c.out from files named a.in and b.in by running the command "./myprog in1 in2 > out" on the remote system.

This required one minor change to the Makeflow syntax.  If you have a variable assignment where the contents of the variable include a ":", then instead of writing it
	URL=http://nd.edu/~ccl/index.html
you now have to write it
	URL:=http://nd.edu/~ccl/index.html

The := indicates that this is explicitly a variable assignment.  All other variable assignments (ones without ":"s in them) work as before.



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@2450 a4d8336d-3463-0410-8bba-c098c45d37a8