To make the time fields reported to catalog server consistent, the unit of WQ master start_time reported to the catalog server is changed to seconds, instead of milliseconds(timestamp_t).
However, the internal representation of master start time (q->start_time) is still in milliseconds, which is the precision we use for various other time variables. And keeping it in milliseconds does not affect the dependent calculations.

Updated the code where the master's start_time field in the catalog server is used.



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@2548 a4d8336d-3463-0410-8bba-c098c45d37a8