Added some comments to struct dag, struct dag_node, and dag_file.
Added a structure (struct dag_parse) to bookkeep tha parsing of a
makeflow file, moving out some of the members of struct dag. This
is to ease the separation between the parsing stage and the dag
logic proper.

Added names to a couple of hard-coded constants
(max_remote_jobs_default, etc.)



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@2578 a4d8336d-3463-0410-8bba-c098c45d37a8