Now the resource_monitor writes the information of all the processes to a single file. A row in the log file is of four types: General information, process runtime information, working directory information, and filesystem information.
Repeated working directories and filesystems are now checked only
once. This is done by keeping reference counts of filesystems and
working directories. Added debug(...) statements to keep track of
reference counts.

Now the processes to be monitored wait for a SIGCONT signal
from the monitor. This allows the monitor to prepare the monitor
structures before a processes start to run. The goal is to extend
this idea to a wrapper of fork.

On SIGINT, the monitor send SIGINT to the first process created,
and does the cleanup of the monitoring tables.

Added hogmem.c, a simple program to test the monitor that
allocates different sizes of memory.



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@2592 a4d8336d-3463-0410-8bba-c098c45d37a8