1. Bug fix: catalog server can only store one worker pool info per host.
The reason is the catalog server use "address-hostname-port" as the key (as in a hashtable) to store updates.
Since a worekr pool is listening on any port, it does not report port info to the catalog.

The solution is to let worker pool report the 'port' item. The value of the worker pool 'port' item is the "MAX_TCP_PORT + worker_pool_pid".  Thus the key (for catalog server lookup) for a worker pool won't conflict with a WQ master's as they have different ranges.

2. Made the worker pool advertisement to the catalog server have 3 min lifetime. And the master and the pool advertisements to the catalog server can have different lifetimes, which can be set in work_queue_catalog.h (WORK_QUEUE_CATALOG_MASTER_AD_LIFETIME and WORK_QUEUE_CATALOG_POOL_AD_LIFETIME).




git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@2608 a4d8336d-3463-0410-8bba-c098c45d37a8