Starts integrating the monitor with makeflow. Adds a flag -M<seconds> to automatically monitor the workflow. A log is produced per rule, and each log is numbered after the rule.
Currently, it is assumed that the executable "resource_monitor"
is in the working directory of the makeflow file (I was trying to
do something more fancy with copying files, but I am getting
errors I do not know how to fix cleanly now).

Incidentally, this makes even clearer the need to separate
lexing, parsing, and dag logic in makeflow.



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@2614 a4d8336d-3463-0410-8bba-c098c45d37a8