Adds the monitoring helper library, that lightly wraps process's related functions, such as fork, wait, open, etc. to monitor all the descendants in a process tree, without something heavy like ptrace, or making assumptions about the system, such as the existence of /proc. The library is loaded automatically by resource_monitor, which launches its first process with LD_PRELOAD.
So far, only fork, vfork, wait and waitpid have been wrapped.

The communication between the library and the monitor is done
with a connectionless socket (datagram socket). The port of the
socket is found automatically, with code very much like the one
used in Work Queue.

The helper library also allows better portability between
systems, since by wrapping chdir, we can monitor changes in the
working directory, without polling /proc/PID/cwd, which does not
exist in all systems. The same follows for io with read and
write.

Once open is wrapped, we can report the files opened during a
workflow, and compare, for example, with the list of files given
by makeflow. We can use this information to report hidden
assumptions in the workflow.

If the helper library cannot be loaded, or the datagram socket
cannot be opened, monitoring follows as usual, but only for the
first process that the monitor spawned.

The library proper is implemented in rmonitor_helper.c. The
communication between the monitor and the library is implemented
in rmonitor_helper_comm.*. The monitor does not depend on the
library, thus the monitor itself can be statically linked.

(Of course, the library is not of much help if the processes to
be monitored are statically linked.)



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@2619 a4d8336d-3463-0410-8bba-c098c45d37a8