Adds termination alarms to the monitor. The user can specify -l maxfile, in which maxfile contains lines such as:
num_processes: 20
memory: 1000000

to indicate the maximum resources allowed. If a resource is not
specified in the file, then it gets the default INTMAX_MAX.

When a resource goes over the limit, the monitor sends SIGTERM to
all the processes (and then SIGKILL to stubborn ones), and in the
summary reports the resource that went over the limit, such as:


command: ./msqrt 100
start:  1361913695666384 Tue Feb 26 16:21:35 2013
end:    1361913698701189 Tue Feb 26 16:21:38 2013
exit-type:      signal 15 Terminated
exit-status:    0
monitor-watch-end: num_processes 46 > 20


This commit also adds msqrt.c, to test cpu-time intensive
processes by trying to compute many simultaneous square roots to
(perhaps unachievable) accuracy.

Finally, it fixes a bug where a file was being left open when
monitoring cpu-time.




git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@2629 a4d8336d-3463-0410-8bba-c098c45d37a8