Generates two versions of the resource_monitor. resource_monitorv (for vanilla) that runs without interacting with the helper library, and resource_monitor which uses the library if it can find it.
Additionally, resource_monitor is now created such that it
contains the helper library. When run, if the helper library does
not exist, resource_monitor writes it first, and then continues
as before. If the library was automatically extracted, then it is
deleted at the end of the monitoring task. This is with the idea
that we need to specify less files, reducing the mistake of
forgetting to include the helper library when rewriting a
makeflow file.



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@2631 a4d8336d-3463-0410-8bba-c098c45d37a8