Makes the file namespace the responsibility of each node. The next step is to generate a single dag_file per filename in the makeflow file.
Adds a parameter to dag_to_file to specify a file renaming
functions. This is particularly needed for the linker. In this
way, collect_input_files can copy the files with the same names
of the newly generated makeflow file.

Given the above, this commit breaks the recently added linker
tests.

The function translate_filename is now
dag_node_translate_filename.



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@2646 a4d8336d-3463-0410-8bba-c098c45d37a8