Every filename in makeflow now gets a unique struct dag_file. The dag_file now also records which nodes have the file as a source, and which node has the file as a target. This simplifies somewhat some of the other code, as it becomes easier to find sources and sinks of the dag. These changes will also make easier to attach attributes to the files, such as needed permissions, size requirements that resource_monitor found, etc.
One major change to note is that dag->file_table now maps all
filenames to dag_files. Before it was a hash table that mapped a
filename to the node that produced it (but this information is
already in dag_file).

I suspect that the functions dag_depth and dag_width now can be
written now more clearly. This is important because there are
some fields in dag_node needed uniquely for these functions,
which makes struct dag_node a little more difficult to read (by a
human, that is).

This changes made the function dag_check_dependencies
unnecessary.



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@2658 a4d8336d-3463-0410-8bba-c098c45d37a8