The monitor now keeps tracks which files has been opened. It writes the list at the end of the log file (seen as comments to gnuplot).
The monitor now also attempts to estimate IO done with files
mapped with mmap. This is done by observing changes in the number
of resident bytes in memory from the map. The method can be quite
inexact, as the monitor is blind to swapping of pages between
intervals. 

A better way would be to count the page faults produced by the
mmap, but Linux does not seem to count file mmap faults as major
page faults (perhaps this information is in /proc/pid/pagemap?).
Another alternative is the value of "read_bytes", from
/proc/pid/io, but this has some issues with distributed
filesystems.



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@2660 a4d8336d-3463-0410-8bba-c098c45d37a8