Monitoring of single tasks in work queue.
If the queue is created with the function
work_queue_create_monitoring (instead of just work_queue_create),
every task submitted is wrapped in the resource_monitor, and a
summary of the resources used is returned. Right now every task
generated a summary file, but this will we consolidated to a
single file in the near future.