change makeflow bundle directory output format
This is the start of the workflow linker.

__Overview__

In this set of changes the linker extracts the dependencies from
Makeflow and Python files.

The linker starts with a Makeflow workflow description and recursively
traverses the dependency tree. Once all of the dependencies are found
the linker creates a packaged version of the original workflow. This
packaged version contains all of the necessary files to run
successfully.