Check that parrot's p->break_address is valid.
For some reason, p->break_address sometimes points outside valid memory maps.
This causes execve to try to execute invalid filenames, generating "Bad
address" errors. Until we figure out why that is, now we check if the address
is valid, and if not, then we find the arguments scratch space with the
fallback function.