Ensures main process gets wait().
When moving from waitid to the current code, if two signals were sent to
the same process, say sigchild and sigterm, sometimes sigterm would kill
the process, with the code for sigchild never running. The problem was
that the call for wait() was only on the handler of sigchild. Before
this was not a problem, since waitid was called on sigchild only for
debugging purposes, with wait() in the handler for sigterm.