Return failure before applying cvmfs_open on a directory path.
If a path refered to a directory, cvmfs lookup tagged the path as correct, but
cvmfs_open would fail with errors "failed to fetch", and "failed to open".
These errors slowed down parrot greatly. Now we check whether lookup tagged the
path as directory, and if so, immediatly return failure to open the path as a
regular file.