Adds libparrot_helper support for 32 and 64 bits.
When LD_PRELOAD contains the token '/$LIB/', this token is substituted by 'lib'
for 32-bit and 'lib64' for 64-bit. This commit tries to compile both the 32 and
64 bit versions of the helper, and installs them respectively in
lib/lib/libparrot_helper.so and lib/lib64/libparrot_helper.so, with
lib/libparrot_helper.so being the default architecture. This allows to write
something like:

setenv PARROT_HELPER '/path/to/cctools/lib/$LIB/libparrot_helper.so'

The single quotes are important, otherwise $LIB gets the value from the shell,
rather than from ld.

The double lib/lib is annoying, but $LIB does not seem to expand when it
is part of name (say '$LIB'parrot_helper.so). This would not be as bad
if $LIB expanded to lib32, rathen than only lib. This commit tries to
preserve the old behaviour when all of the execs run in the default
architecture.

When parrot detects the $LIB token, it blindly assumes that the path to
the libraries is correct. Perhaps a better way is to expand $LIB
ourselves to test the paths.

Also, when the configure script cannot compile one of the test versions, the
errors are sent to /dev/null. A better way would be to send them to a
log, or print them nicely to stdout, so a user can fix them.

(I tested this on a fresh install of SL5.5. For the 32-bit version we seem to
be missing the i386 dev package in our installation of redhat5.5.)