cores/memory/disk requirements in makeflow/condor
Examine BATCH_OPTIONS to see if Requirements has been defined. If not,
add the resources the easy way. Otherwise, we need to look for the
original requirements and conjunct them with the cores/memory/disk
specified. A little hackish, but it kind of works...

For example, if:
export CORES=3
export MEMORY=10000
export BATCH_OPTIONS="Requirements = Disk >= 10000"

we end up with a condor.submit with a line:

Requirements = ( Disk >= 10000) && ((Cores>=3) && (Memory>=10000))