Fixed double-free affecting shrimp workflow.
The problem arose when a worker would be flagged for both not having sent its ready message in time and not having responded to a keepalive in time.  Each condition causes the worker to be removed, and the second removal resulted in a double free.

We now no longer send keepalives to non-ready workers, since they don't have any tasks or data assigned to them being held "hostage", and hung workers are handled by the new code.