Fixes issue with foreman and unlabled tasks.
Any work_queue master assigning a completely unlabled task to its subordinate assumes the task will consume the average resources advertised by that entity.

Tasks that only have one or two resources labled are assumed to "take whatever is left" of the unlabeled components, and are marked as consuming zero.