Move path resolution to CFS layer.
The Chirp server now only URL decodes and collapses paths from the wire. The
CFS layer now resolves these paths which usually involves prepending the root
path and a final collapse.

The net result of this change is reduced coupling of modules. Now every module
"above" the CFS layer sees the same abstraction of the Chirp root as clients.