Introduces work_queue_submit_internal for foremen.
This commit fixes a bug in which work_queue_submit increased
automatically the taskid in a foreman.  This is incorrect for foreman,
where taskid are set by the master. This commit makes work_queue_submit
a wrapper for a new function, work_queue_submit_internal, which submit
tasks without modifying the taskid.

Without this fix, the taskid at the master and the foreman would be
different. When the foreman reported the task as completed, two problems
were caused:

1) If the master never had such a taskid, it would just ignore the
results from the foreman.

2) Even worse, if the foreman taskid did exits at the master side, the
master would ask to transfer files that did not exist on the foreman.