Enable usleep in sqlite for cheaper timeouts.
Early on, chirp_job relied on sqlite3_busy_timeout to reduce EAGAIN errors to
the client. Unfortunately, it didn't behave as expected and would take
mysteriously long amounts of time to acquire locks. So, I changed the code to
usleep(2000) and then manually restart chirp_job RPCs when the database was
locked.

However, I recently noticed we simply need to set HAVE_USLEEP when compiling
sqlite3.c. This allows us to use sqlite3_busy_timeout again which
reduces/eliminates the "database is locked" messages we see.

This commit also disables pthread_mutex_* protections in sqlite3 which are
unnecessary for Chirp (as it is single-threaded).