Add chirp_fuse_create method.
FreeBSD 10, recentely released, now officially supports fuse.

On Linux, if the create method is not implemented, then mknod and open
are called instead. This is not true in FreeBSD, where trying to open a
new file results in "Operation not implemented". This commit implements
create in terms of mknod and open, preserving the old behaviour in
Linux.