Set glibc variables tracking the program name.
This also affects /proc/pid/cmdline.

I think another way to do this is Linux prctl and POSIX pthread_setname_np
(which on Linux wraps prctl). However, there are two problems: (1) prctl limits
the length to 16 characters including the NUL terminator and (2) not all
utilities read /proc/pid/comm (which prctl changes). It seems most utilities
prefer /proc/pid/cmdline which is based off of argv[0]. How it is actually
changed is not very clear. These glibc variables appear to affect it.

setproctitle (BSD, but also on Linux) is another way to change the title.