CVMFS thread hackish bugfix.
As the comment in the source:

There is an unidentified race condition when using threads and cvmfs, in
which the thread misses the SIGCONT. Sleeping for a second seems to give
a chance to the thread to catch the signal.  Hopefully we will be able
to remove this hack soon.