Correct WQ binding build error due to externals.
ND's iRODS is compiled without -fPIC so the linker terminates with an error
when trying to build the WQ binding shared libraries. This fix is a stopgap to
get the build working: we separate out the archives and shared libraries we
link against and only link against shared libraries for the WQ bindings.  Since
none of the WQ bindings depend on any external archives, this solves the
problem.

Ideally, we would separate out each individual external and dependency so we
could link only against what we want. However, that would be a much larger and
intrusive change to the build system.

Fixes #466.